refactor(domain): extract InMemoryStore into kanban-backend-memory
Moves the in_memory_store module out of kanban-domain into a new
kanban-backend-memory crate, together with its KanbanBackend impl, so
the in-memory backend is a peer of json/sqlite/http rather than a
privileged implementation living inside the domain crate.

kanban-domain now holds only entities, commands, and trait definitions;
kanban-backend holds only the abstractions, with no impl and no
test-helpers feature.

MockBackend moves with it. It delegates to an inner InMemoryStore, so
leaving it in kanban-backend would make that crate depend on
kanban-backend-memory, which depends back on kanban-backend. A
dev-dependency does not break the cycle because the module compiled into
the lib under its feature gate. It now lives in kanban-service, its only
consumer, under cfg(test).

The nine KanbanBackend lifecycle tests move to kanban-backend-memory
alongside the impl they exercise.
