+++ crates/kanban-domain/tests/common/mod.rs
@@ -2,8 +2,8 @@
 //! subset, so unused items would otherwise warn as dead code per-binary.
 #![allow(dead_code)]
 
+use kanban_backend_memory::InMemoryStore;
 use kanban_domain::commands::CommandContext;
-use kanban_domain::InMemoryStore;
 
 pub struct TestContext {
     pub store: InMemoryStore,
