+++ crates/kanban-persistence/Cargo.toml
@@ -41,6 +41,7 @@ tracing.workspace = true
 tempfile = { workspace = true, optional = true }
 
 [dev-dependencies]
+kanban-backend-memory = { path = "../kanban-backend-memory" }
 mockall.workspace = true
 tempfile.workspace = true
 tokio = { workspace = true, features = ["full"] }
