+++ crates/kanban-service/Cargo.toml
@@ -24,6 +24,7 @@ kanban-domain = { path = "../kanban-domain", version = "^0.7" }
 kanban-persistence = { path = "../kanban-persistence", version = "^0.7" }
 kanban-api = { path = "../kanban-api", version = "^0.7" }
 kanban-backend = { path = "../kanban-backend", version = "^0.7" }
+kanban-backend-memory = { path = "../kanban-backend-memory", version = "^0.7" }
 kanban-persistence-json = { path = "../kanban-persistence-json", version = "^0.7", optional = true }
 kanban-persistence-sqlite = { path = "../kanban-persistence-sqlite", version = "^0.7", optional = true }
 async-trait.workspace = true
@@ -40,7 +41,6 @@ tracing.workspace = true
 tempfile.workspace = true
 
 [dev-dependencies]
-kanban-backend = { path = "../kanban-backend", features = ["test-helpers"] }
 kanban-persistence-json = { path = "../kanban-persistence-json" }
 kanban-persistence-sqlite = { path = "../kanban-persistence-sqlite", features = ["test-helpers"] }
 kanban-persistence = { path = "../kanban-persistence", features = ["test-helpers"] }
