+++ crates/kanban-service/tests/completion_sync.rs
@@ -1,7 +1,8 @@
 //! KAN-394: status ↔ completion-column auto-sync orchestrated at the service layer.
 
+use kanban_backend_memory::InMemoryStore;
 use kanban_domain::{
-    BoardUpdate, CardStatus, CardUpdate, FieldUpdate, InMemoryStore, KanbanOperations, KanbanResult,
+    BoardUpdate, CardStatus, CardUpdate, FieldUpdate, KanbanOperations, KanbanResult,
 };
 use kanban_service::KanbanContext;
 use std::sync::Arc;
