+++ crates/kanban-tui/tests/data_integrity_tests.rs
@@ -1,3 +1,4 @@
+use kanban_backend_memory::InMemoryStore;
 use kanban_domain::commands::*;
 use kanban_domain::*;
